% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalHelpers.R
\name{abbrevSiteName}
\alias{abbrevSiteName}
\title{Abbreviate a site name}
\usage{
abbrevSiteName(x)
}
\arguments{
\item{x}{a vector of site names}
}
\value{
a vector with all site names abbreviated
}
\description{
Create standard abbreviations for remnant populations
}
