% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpsFunctions.R
\name{addToSurv.csv}
\alias{addToSurv.csv}
\title{Add files for surv.csv}
\usage{
addToSurv.csv(dataFileName, folder = folderRD)
}
\arguments{
\item{dataFileName}{the name of the RData file in the standard format.
Adding ".RData" is not necessary.}

\item{folder}{the folder where the RData file can be found}
}
\description{
Add all points from the RData file into surv.csv. If surv.csv doesn't exist,
it will be created. Currently it gets saved in the working directory but
it could be changed to a standard dropbox location.
}
\keyword{gps}

