% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialFunctions.R
\name{assignEchinaceaSite}
\alias{assignEchinaceaSite}
\title{Function assignEchinaceaSite:}
\usage{
assignEchinaceaSite(df)
}
\arguments{
\item{df}{data frame with Ax and Ly columns representing MN state plane coordinates}
}
\value{
vector of standardized site names
}
\description{
Function to automatically assign standardized Echinacea sites based on spatial coordinates
}
\details{
assign standardized site names based on spatial coordinates
}
\author{
Jared Beck
}
