% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpsFunctions.R
\name{backupGPSFiles}
\alias{backupGPSFiles}
\title{Transfer files from GPS to geospatialDataBackup folder}
\usage{
backupGPSFiles(machine, year, user = "Echinacea")
}
\arguments{
\item{machine}{which GPS machine (sulu or chek) that is being tranferred from}

\item{year}{the year that the GPS points were taken}

\item{user}{rarely change this but if using a computer other than Hal, change
so that the file path will work (e.g. "dhanson")}
}
\value{
The names of the files that were transferred from the GPS to
 geospatialDataBackup.
}
\description{
Find all job that are on the connected GPS unit and move to geospatialDataBackup.
It only works on Hal (and probably R2D2) unless your computer can connect to the GPS
}
\keyword{gps}

