% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageCountData.R
\name{batchSummary}
\alias{batchSummary}
\title{The batch summary function summarizes information from harvest datasheet (hh.year)}
\usage{
batchSummary(hh)
}
\arguments{
\item{hh}{harvest data frame (e.g. hh.2014)}
}
\value{
summary, a summary of each cg batch including letNo range and the number of heads
}
\description{
The batch summary function summarizes information from harvest datasheet (hh.year)
}
