% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageScans.R
\name{check.year}
\alias{check.year}
\title{Compare letnos or nolets from scan files with harvest list for a given 
year.}
\usage{
check.year(scanFolder, harvestFile, writeTo = "./scanSummary.csv")
}
\arguments{
\item{scanFolder}{the location of the scan files}

\item{harvestFile}{one of the hh.year files that comes with this package}

\item{writeTo}{the file to which the summary will be written}
}
\value{
none, write to a csv
}
\description{
This function runs check.batch over all experiments and then creates a
summary for what is missing.
}
