% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massFunctions.R
\name{combineMassFiles}
\alias{combineMassFiles}
\title{combine raw mas data txt file into useful output}
\usage{
combineMassFiles(path = ".", writeCsv = FALSE,
  fileName = "allMassFiles.csv")
}
\arguments{
\item{path}{character designating directory containing files of interest}

\item{writeCsv}{logical indicating whether to print output to csv file}

\item{fileName}{character for name of output file}
}
\value{
dataframe of mass data from all raw mass files in directory of 
  interest, returned invisibly. A csv file is returned if writeCsv is TRUE.
}
\description{
This function combines all good mass records in a single dataframe and, optionally,
writes a csv.
}
\examples{

\dontrun{
setwd("I:\\\\Departments\\\\Research\\\\EchinaceaVolunteers\\\\Balance\\\\sampleForEchLab\\\\CG2009_goodFiles")
tt <- combineMassFiles()
}
}
\seealso{
\code{\link{listBadFiles}} and \code{\link{readMassFile}} and 
  \code{\link{investigateMassFiles}} which are other useful functions that 
  deal with mass files
}
\keyword{file}
\keyword{mass}

