% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalHelpers.R
\name{compare2vectors}
\alias{compare2vectors}
\title{Compare the contents of two vectors}
\usage{
compare2vectors(v1, v2, showInBoth = FALSE)
}
\arguments{
\item{v1}{the first vector to compare}

\item{v2}{the second vector to compare}

\item{showInBoth}{whether or not to display the values found in
both vectors. Default is FALSE}
}
\value{
a list containing (1) a data frame with various counts (2) a vector
with elements only in the first vector (3) a vector with elements only in
the second vector (4) if showInBoth == TRUE, a vector with elements found
in both vectors
}
\description{
This function takes two vectors and displays how many elements are
in one, in the other, or in both as well as which ones are only
in one of the two vectors.
}

