% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsFunctions.R
\name{convertXMLtoRdata}
\alias{convertXMLtoRdata}
\title{Convert an XML file to RData}
\usage{
convertXMLtoRdata(fn, readPath = folderXML, writePath = folderRD)
}
\arguments{
\item{fn}{name of the XML file to convert}

\item{readPath}{file path where the XML file can be found}

\item{writePath}{file path to write the RData file to}
}
\value{
The name of the converted file (without the path)
}
\description{
Converts an XML file into a data frame that is stored in an RData file.
Only the information that we want is taken from the XML
}
\keyword{gps}

