% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageCountData.R
\name{createCSV}
\alias{createCSV}
\title{create assignment csv to upload}
\usage{
createCSV(scansdf = scans, harvYear, priority = 50, type = c("imageTool",
  "online", "xray"))
}
\arguments{
\item{scansdf}{is a data frame of scan filenames and letnos, scans (from function loadScans) by default}

\item{harvYear}{is the year heads were harvested, this needs to be entered}

\item{priority}{is the counting priority of the scans, 50 by default}

\item{type}{the type of upload csv to make, imageTool has filepaths starting
with "C:/", online has file paths starting with the online achene counter, and xray
has file paths with the online x-ray counter}
}
\value{
returns out, an upload-ready data.frame

simply type in write.csv(out, "C:/assignmentYear.csv", row.names = FALSE to save)
}
\description{
createCSV automates the creation of an upload-ready csv for the Echinacea Project's online counting database.
This function now works for achene counts and x-rays.
The function is set up to create 3 records for each letno
(3 rounds of counting). The function randomizes the order of records,
no need for pre-upload randomization
}
