% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gpsFunctions.R
\name{errorCheckRData}
\alias{errorCheckRData}
\title{Check for GPS errors}
\usage{
errorCheckRData(dataFileName, folder = folderRD)
}
\arguments{
\item{dataFileName}{the name of the RData file in the standard format.
Adding ".RData" is not necessary.}

\item{folder}{the folder where the RData file can be found}
}
\description{
Check for too high standard deviation in GPS points and print out
problem points. Also print out any survNote's.
}
\keyword{gps}

