% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aiiFunctions.R
\name{estimateSeedSet}
\alias{estimateSeedSet}
\title{Estimate seed set in a head}
\usage{
estimateSeedSet(df, partials = TRUE, idCol = "headID",
  totalCols = c("topCount", "middleCount", "bottomCount"),
  fullCols = c("topFull", "middleFull", "bottomFull"),
  sampCols = c("topCount", "informativeCount", "bottomCount"),
  partialCols = c("topPartial", "middlePartial", "bottomPartial"))
}
\arguments{
\item{df}{a data frame that contains seed set data}

\item{partials}{if TRUE, partials will be included in}

\item{idCol}{the column that contains the ids for each head}

\item{totalCols}{a vector with names of the columns that contain counts for
the total number of achenes in the top, middle, and bottom of the head}

\item{fullCols}{a vector with names of the columns that contain counts for
the number of full achenes in the top, middle, and bottom of the head}

\item{sampCols}{a vector with names of the columns that contain counts for
the number of achenes in the x-rayed sample for the top, middle, and bottom
of the head}

\item{partialCols}{a vector with names of the columns that contain counts for
the number of partial achenes in the top, middle, and bottom of the head. If
\code{partial} is FALSE, this can be ignored}
}
\value{
a data frame containing the id and seed set data for two methods
of calculating it. Columns "pointEstFull" and "acheneCt" correspond to the
point estimate method and columns "sampleFull" and "sampleCount" correspond
to the resampling method.
}
\description{
Estimate seed set in a head that was split into top, middle, and bottom
using two different methods: one resampling based, the other a point estimate
}

