% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cgFunctions.R
\name{exPt01CoreDataSummary}
\alias{exPt01CoreDataSummary}
\title{Generate summary and upload status of exPt01 core dataset}
\usage{
exPt01CoreDataSummary(df = NULL, targetYear = NA, expNames = NA)
}
\arguments{
\item{df}{alternative exPt01 core dataframe in wide format. It must have columns in this format: 
cgPlaId, yrPlanted, ld20xx, fl20xx, hdCt20xx, achCt20xx. All fields must be of "integer" or "numeric" class.
Default is the core dataset online: "http://echinaceaproject.org/data/cg1CoreData.csv"}

\item{targetYear}{any year past 1996. Data summary will be displayed for each experiment through your target year,
if the appropriate data exists. If not, summaries will default to most recent year with complete data for 
each experiment. The status matrix is unaffected by this argumant, and will always display the most recent 
available data for each experiment and field type. Default is most recent data.}

\item{expNames}{vector of experiment names you want a status and a summary for. Possible experiments are:
c("1996", "1997", "1998", "1999", "1999S", "2001", "INB2", "INB1", "M03", "qGen", "SPP"). Note that you may also
use "Inbreeding" for "INB1" and "Monica 2003" for "M03". Matrices will still display "INB1" and "M03". Default
is all exPt01 experiments.}
}
\description{
This function will read the cg1Core dataset from echinaceaproject.org and generate a matrix of how up-to-date
the cg1Core dataset for each experiment. Additionally, this function will generate a matrix of summary data
for the exPt01 core dataste, including ld, fl, hdCt, and achCt. Finally, this function will warn users of
missing data and NA's in the core dataset. You may use a different core ExPt01 dataset, input a year you 
would like a summary of exPt01 through, and input a vector of the specific experiments you would like to 
learn about.
}
\details{
use this function to get a summary of the core exPt01 dataset, learn where data must still be updated,
and learn of any discrepancies in the data. Function will output summary matrices p1Status and p1Summary. achCt for
certain experiments may be behind others because for certain years, they may have hdCt > 0 for a cgPlaId, but an
achCt of 0. This indicates missing data, and the function will not consider it to be "up-to-date" data. The reason 
df = NULL and the other two parameters = NA is because sw and rt are different and sometimes different is fun...ction.
}
\examples{
\dontrun{exPt01CoreDataSummary()}

\dontrun{exPt01CoreDataSummary(cc, 2002, c("M03", "INB2", "INB1", "1996"))}

\dontrun{exPt01CoreDataSummary(targetYear = 2001)}



}
