% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generalHelpers.R
\name{getHeadIDPiece}
\alias{getHeadIDPiece}
\title{Get one piece of the headID}
\usage{
getHeadIDPiece(headIDs, piece = c("tag", "tt", "year"))
}
\arguments{
\item{headIDs}{character vector of head ids in the form TAG-TT-YEAR}

\item{piece}{character. One of "tag" if you want the tag, "tt" if you want
the twist tie, or "year" if you want the year}
}
\value{
a character vector with the piece that you want
}
\description{
Get one piece of the headID
}

