\docType{data}
\name{hh.2013}
\alias{hh.2013}
\title{Information about heads harvested from cg1 in 2013.}
\format{A 2336 x 10 data frame}
\usage{
hh.2013
}
\description{
This data frame contains harvest information on all heads
harvested from cg1 in 2013. There are a total of 2336
records. Only certain heads are included. This dataset was
generated by the script make_hh.2013.r (This file was generated by the script make_hh.2013.r, which is save in ../Dropbox/CGData/145_labelEnvelopes).
}
\section{Variables}{
  Variables: \itemize{ 
  \item cgHdId, unique identifier for heads 
  \item batch, this is the experiment name 
  \item block, Lydia doesn't know what this is
  \item lineNo, line on the paper datasheet 
  \item letno, unique identifier for heads  
  \item No, the four digit unique identifier
  \item row, emmr/1000 
  \item pos, nmmp/1000 
  \item tt, twist-tie color
  \item gBag, the grocery bag that the heads was harvested }
}
\examples{
dim(hh.2013)
str(hh.2013) # we should include useful example code
}
\references{
we need some more info
}
\keyword{datasets}

