\docType{data}
\name{hh.2014}
\alias{hh.2014}
\title{Information about heads harvested from p1 and p2 in 2014.}
\format{A 811 x 10 data frame}
\usage{
hh.2014
}
\description{
This data frame contains harvest information on all heads
harvested from p1 and p2 in 2014. There are a total of 811
records. This data frame was generated by the script 
make.hh.2014.rda.r, which is saved in 
..Dropbox/CGData/140_reconcile/reconcile2014/reconcile2014_out).
}

\section{Variables}{
  Variables: \itemize{ 
  \item cgHdId, unique identifier for heads 
  \item batch, this is the experiment name 
  \item block, identifiers harvesting area
  \item lineNo, line on the paper datasheet 
  \item letno, unique identifier for heads  
  \item No, the four digit unique identifier
  \item row, emmr/1000 
  \item pos, nmmp/1000 
  \item tt, twist-tie color
  \item gBag, the grocery bag in which harvester placed head }
}
\examples{
dim(hh.2014)
str(hh.2014) # we should include useful example code
}
\references{
we need some more info
}
\keyword{datasets}

