\docType{data}
\name{hh.2015}
\alias{hh.2015}
\title{Information about heads harvested from p1 and p2 in 2015.}
\format{A 3223 x 10 data frame}
\usage{
  hh.2015
}
\description{
  This data frame contains harvest information on all heads
  harvested from p1 and p2 in 2015. There are a total of 3223
  records. This data frame was generated by the script 
  make.hh.2015.r, which is saved in 
  ..Dropbox/CGData/140_reconcile/reconcile2015/reconcile2015_out.
  There were 9 records that didn't have gbags listed because they
  weren't harvested. Those can be found by searching through 
  Dropbox/CGData/140_reconcile/reconcile2015/reconcile2015_out/reconciledCGHarvest2015.2016Jan25.csv
  If they get figured out, add them to hh.2015.rda
}

\section{Variables}{
  Variables: \itemize{ 
    \item cgHdId, unique identifier for heads 
    \item batch, this is the experiment name 
    \item block, not sure so everything is 0
    \item lineNo, the order by row number then position number 
    \item letno, unique identifier for heads  
    \item No, the four digit unique identifier
    \item row, emmr/1000 
    \item pos, nmmp/1000 
    \item tt, twist-tie color
    \item gBag, the grocery bag in which harvester placed head }
}
\examples{
  dim(hh.2015)
  str(hh.2015) # we should include useful example code
}
\keyword{datasets}

