\docType{data}
\name{hh.2016}
\alias{hh.2016}
\title{Information about heads harvested from p1 and p2 in 2016.}
\format{A 1065 x 10 data frame}
\usage{hh.2016}
\description{
  This data frame contains harvest information on all heads
  harvested from p1 and p2 in 2016. There are a total of 1065
  records. This data frame was generated by the script 
  make.hh.2016.r, which is saved in 
  ..Dropbox/CGData/140_reconcile/reconcile2016/reconcile2016_out.
  The records come from the reconciled harvest list in the same
  folder as above.
}
\section{Variables}{
  Variables: \itemize{ 
    \item gBag, the grocery bag in which harvester placed head
    \item row, row of common garden head was collected from
    \item tt, twist-tie color 
    \item cgHdId, unique identifier for head 
    \item batch, experiment name for head
    \item block, ?
    \item lineNo, line of paper harvest datasheet for head
    \item No, the four digit unique identifier
    \item pos, position in row of common garden of head
    \item letno, unique identifier for heads  }
}
\examples{
  dim(hh.2016)
  str(hh.2016)
  
  table(hh.2016$batch)
  
  plot(pos ~ row, hh.2016[hh.2016$pos > 800,]) # plot of heads harvested in p1 & 99s
}

\keyword{datasets}