\docType{data}
\name{hh.2017}
\alias{hh.2017}
\title{Information about heads harvested from p1 and p2 in 2017.}
\format{A 1150 x 10 data frame}
\usage{hh.2017}
\description{
  This data frame contains harvest information on all heads
  harvested from p1 and p2 in 2017. There are a total of 1150
  records. This data frame was generated by the script 
  make.hh.2017.r, which is saved in 
  ..Dropbox/CGData/140_reconcile/reconcile2017/reconcile2017_work.
  The records come from the reconciled harvest list in 
  ..Dropbox/CGData/140_reconcile/reconcile2017/reconcile2017_out.
}
\section{Variables}{
  Variables: \itemize{ 
    \item gBag, the grocery bag in which harvester placed head
    \item row, row of common garden head was collected from
    \item tt, twist-tie color 
    \item cgHdId, unique identifier for head 
    \item batch, experiment name for head
    \item block, we used to use this to prioritize cleaning within experiments
    \item lineNo, line of paper harvest datasheet for head
    \item No, the four digit unique identifier
    \item pos, position in row of common garden of head
    \item letno, unique identifier for heads  }
}
\examples{
  dim(hh.2017)
  str(hh.2017)
  
  table(hh.2017$batch)
  
  plot(pos ~ row, hh.2017[hh.2017$pos > 800,], asp = 1) # plot of heads harvested in p1 & 99s
  plot(pos ~ row, hh.2017[hh.2017$pos < 800,], asp = 1) # plot of heads harvested in p2
}

\keyword{datasets}