\docType{data}
\name{hh.2018}
\alias{hh.2018}
\title{Information about heads harvested from p1, p2, and p7 in 2018.}
\format{A 2011 x 10 data frame}
\usage{hh.2018}
\description{
  This data frame contains harvest information on all heads
  harvested from p1, p2, and p7 in 2018. There are a total of 2011
  records. This data frame was generated by the script 
  make.hh.2018.r, which is saved in 
  ..Dropbox/CGData/140_reconcile/reconcile2018/reconcile2018_work.
  The records come from the reconciled harvest list in 
  ..Dropbox/CGData/140_reconcile/reconcile2018/reconcile2018_work.
}
\section{Variables}{
  Variables: \itemize{ 
    \item gBag, the grocery bag in which harvester placed head
    \item row, row of common garden head was collected from
    \item tt, twist-tie color 
    \item cgHdId, unique identifier for head 
    \item batch, experiment name for head
    \item block, we used to use this to prioritize cleaning within experiments
    \item lineNo, line of paper harvest datasheet for head
    \item No, the four digit unique identifier
    \item pos, position in row of common garden of head
    \item letno, unique identifier for heads  }
}
\examples{
  dim(hh.2018)
  str(hh.2018)
  
  table(hh.2018$batch)
  
  plot(pos ~ row, hh.2018[hh.2018$pos > 800,], asp = 1) # plot of heads harvested in p1 & 99s
  plot(pos ~ row, hh.2018[hh.2018$pos < 800,], asp = 1) # plot of heads harvested in p2 & p7
}

\keyword{datasets}