\docType{data}
\name{hh.2019}
\alias{hh.2019}
\title{Information about heads harvested from p1, p2, p8, and p9 in 2019.}
\format{A 1186 x 10 data frame}
\usage{hh.2019}
\description{
  This data frame contains harvest information on all heads
  harvested from p1, p2, p8, and p9 in 2019. There are a total of 1186
  records. Included is head CT-2036, which was not found during reconcile but was recorded as harvested. 
  This data frame was generated by the script 
  make.hh.2019.r, which is saved in 
  ..Dropbox/CGData/140_reconcile/reconcile2019/reconcile2019_work.
  The records come from the reconciled harvest list in 
  ..Dropbox/CGData/140_reconcile/reconcile2019/reconcile2019_out.
}
\section{Variables}{
  Variables: \itemize{ 
    \item gBag, the grocery bag in which harvester placed head
    \item row, row of common garden head was collected from
    \item tt, twist-tie color 
    \item cgHdId, unique identifier for head 
    \item batch, experiment name for head
    \item block, we used to use this to prioritize cleaning within experiments
    \item lineNo, line of paper harvest datasheet for head
    \item No, the four digit unique identifier
    \item pos, position in row of common garden of head
    \item letno, unique identifier for heads  }
}
\examples{
  dim(hh.2019)
  str(hh.2019)
  
  table(hh.2019$batch)
  
  plot(pos ~ row, hh.2019[hh.2019$pos > 800,], asp = 1) # plot of heads harvested in p1 & 99s
  plot(pos ~ row, hh.2019[hh.2019$pos < 800,], asp = 1) # plot of heads harvested in p2, p8, and p9
}

\keyword{datasets}