\docType{data}
\name{hh.2020}
\alias{hh.2020}
\title{Information about heads harvested from p1, p2, p5, p7, p8, and p9 in 2020.}
\format{A 1419 x 10 data frame}
\usage{hh.2020}
\description{
  This data frame contains harvest information on all heads
  harvested from p1, p2, p5, p7, p8, and p9 in 2020. There are a total of 1419
  records. 
  This data frame was generated by the script 
  makehh2020.r, which is saved in 
  ..Dropbox/CGData/140_reconcile/reconcile2020/reconcile2020_work.
  The records come from the reconciled harvest list in 
  ..Dropbox/CGData/140_reconcile/reconcile2020/reconcile2020_out.
}
\section{Variables}{
  Variables: \itemize{ 
    \item gBag, the grocery bag in which harvester placed head
    \item row, row of common garden head was collected from
    \item tt, twist-tie color 
    \item cgHdId, unique identifier for head 
    \item batch, experiment name for head
    \item block, we used to use this to prioritize cleaning within experiments
    \item lineNo, line of paper harvest datasheet for head
    \item No, the four digit unique identifier
    \item pos, position in row of common garden of head
    \item letno, unique identifier for heads  }
}
\examples{
  dim(hh.2020)
  str(hh.2020)
  
  table(hh.2020$batch)
  
  plot(pos ~ row, hh.2020[hh.2020$pos > 800,], asp = 1) # plot of heads harvested in p1 & 99s
  plot(pos ~ row, hh.2020[hh.2020$pos < 800 & hh.2020$pos > 600 , ], asp = 1) # plot of heads harvested in p8 
  plot(pos ~ row, hh.2020[hh.2020$pos < 600, ], asp = 1) # plot of heads harvested in p2 and p9

}

\keyword{datasets}