% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massFunctions.R
\name{investigateMassFiles}
\alias{investigateMassFiles}
\title{investigate all txt files in a directory}
\usage{
investigateMassFiles(path = ".")
}
\arguments{
\item{path}{character designating directory containing files of interest}
}
\value{
data frame that summarizes characteristics of mass files
}
\description{
Investigate the raw mass files in a directory to find potential errors.
}
\examples{

\dontrun{
setwd("I:\\\\Departments\\\\Research\\\\EchinaceaVolunteers\\\\Balance\\\\sampleForEchLab\\\\CG2009_rawFiles")
investigateMassFiles()
}

}
\seealso{
\code{\link{combineMassFiles}} and \code{\link{readMassFile}} and
  \code{\link{listBadFiles}} which are other useful functions that
  deal with mass files
}
\keyword{file}
\keyword{mass}
