% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialFunctions.R
\name{is.stakefile}
\alias{is.stakefile}
\title{Function is.stakefile: utility function for}
\usage{
is.stakefile(df)
}
\arguments{
\item{df}{data frame with potential stakefile}
}
\value{
TRUE or FALSE
}
\description{
Function to assess whether data frame conforms to stakefile formatting
}
\details{
check formatting of stakefiles
}
