% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialFunctions.R
\name{latlong2stplane}
\alias{latlong2stplane}
\title{Function latlong2stplane: utility function to}
\usage{
latlong2stplane(df)
}
\arguments{
\item{df}{data frame with columns for latitude longitude coordinates}
}
\value{
Data frame with original state plane coordinates as well as lat/long coordinates
}
\description{
Function to convert latitude/longitude points to MN state plane coordinates
}
\details{
convert MN state plane coordinates to lat/long
}
\author{
Jared Beck
}
