% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpsFunctions.R
\name{makeGPSFileName}
\alias{makeGPSFileName}
\title{Make standard GPS job name}
\usage{
makeGPSFileName(fileName)
}
\arguments{
\item{fileName}{the name of the file for which we want a standard name}
}
\value{
The new file name
}
\description{
Takes the file name and converts it to a standard format
}
\keyword{gps}

