% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageXray.R
\name{makeXrayDf}
\alias{makeXrayDf}
\title{Make x-ray data frame}
\usage{
makeXrayDf(hh, batchSize = 200, sheetSize = 20)
}
\arguments{
\item{batchSize}{the number of heads to fit in each subbatch. Default is 200.}

\item{sheetSize}{the number of heads to fit on each x-ray grid sheet}
}
\value{
a data frame with five columns: letno, batch (i.e., experiment), 
No (the numeric part of a letno), subbatch, sheetNo
}
\description{
This function takes in a harvest data frame (e.g., hh.2016) and assigns
a sheet number and subbatch to each head. Subbatches are smaller subunits
of large batches (experiments) for ease of processing.
}
\details{
Subbatches are mainly useful for breaking large batches into smaller,
more manageable chunks. If the number of heads is larger than the size specified
for subbatches (i.e., the argument 'batchSize'), then the subbatch field will take
the format "experimentName_subbatchNumber', where the subbatchNumber is the first
No (numeric part of letno) in that batch.
}
\examples{
xr16 = makeXrayDf(hh.2016, batchSize = 100)
head(xr16)
tail(xr16)
table(xr16$subbatch)

}
\keyword{df}
\keyword{x-ray}
\keyword{xray}
