% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageCountData.R
\name{manageCountData}
\alias{manageCountData}
\title{This function can be used to check count data for errors}
\usage{
manageCountData(df, viewAllMismatches = FALSE)
}
\arguments{
\item{df}{data.frame of count data. This dataframe will be the csv file downloaded from the Ech. Project website}

\item{viewAllMismatches}{logical default is FALSE}
}
\value{
output, a list containing 5 objects: mismatches, cantCount, checkNotes, missingCounts, and checkZeros

mismatches, counter noted that the envelope letNo in the scan did not match the computer letNo

cantCount, counter noted that the scan could not be counted

checkNotes, records where counter made a note. Check through notes!

missingCounts, letNos that were not counted 3 times

checkZeros,
}
\description{
This function can be used to check count data for errors
}
