% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialFunctions.R
\name{siteNames}
\alias{siteNames}
\title{list standard site names ####}
\usage{
siteNames()
}
\value{
vector of standard site names
}
\description{
return vector of standard Echinacea site names
}
\author{
Stuart
}
