% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialFunctions.R
\name{stakefile2avenza}
\alias{stakefile2avenza}
\title{Function stakefile2avenza: utility function}
\usage{
stakefile2avenza(df, fpath)
}
\arguments{
\item{df}{data frame representing stakefile with MN state plane coordinates}

\item{fpath}{file path for exporting kml file}
}
\value{
writes kml file to specified filepath
}
\description{
Function to convert stakefile into kml file for Avenza
}
\details{
convert stakefile into kml file that can be imported into Avenza
}
\author{
Jared Beck
}
