% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geospatialFunctions.R
\name{stplane2latlong}
\alias{stplane2latlong}
\title{Function stplane2latlong: utility function to}
\usage{
stplane2latlong(df)
}
\arguments{
\item{df}{data frame with columns for MN state plane coordinates coded as Ax or x and Ly or y}
}
\value{
Data frame with original state plane coordinates as well as lat/long coordinates
}
\description{
Function to convert MN state plane coordinates in data frame to latitude/longitude
}
\details{
convert MN state plane coordinates to lat/long
}
\author{
Jared Beck
}
