% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadScans.r
\name{transferAllFiles}
\alias{transferAllFiles}
\title{Transfer files from all subfolders to analogous subfolders}
\usage{
transferAllFiles(from, to, askBeforeContinue = TRUE)
}
\arguments{
\item{from}{path to a folder whose subfolders you want checked}

\item{to}{the drive on which the wanted folder is found (e.g. "F:/" or "D:/")}

\item{askBeforeContinue}{logical: should a message be printed before continuing
to check other subfolders}
}
\description{
This function will user the transferFiles function to go through all
subfolders of the selected folder and transfer files not in one folder
to the other
}
\examples{
\dontrun{
path1 <- "C:/cg2014scans"
pathos <- "E:/"
transferAllFiles(pathos, path)}
}
\keyword{scans}

