% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageScans.R
\name{transferFiles}
\alias{transferFiles}
\title{Transfer files from one folder to another}
\usage{
transferFiles(from, to, showNotInFrom = FALSE)
}
\arguments{
\item{from}{path to a folder that you want all files from}

\item{to}{path to a folder which you want files transferred to}

\item{showNotInFrom}{logical: should files not in the from folder but in
the to folder be printed}
}
\description{
This function will compare the contents of two folders and transfer
any files that are not in the second folder but are in the first folder
from the first to the second. If the second folder doesn't exist, it will
create one. It will also inform you how many files are
in the second folder but not the first
}
\examples{
\dontrun{
path <- "C:\\\\Users\\\\dhanson\\\\Documents\\\\scanTest"
pathos <- "E:\\\\cg2014scans\\\\exPt2"
transferFiles(pathos, path, showNotInFrom = TRUE)}

}
\keyword{scans}
