% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageScans.R
\name{transferScanFiles}
\alias{transferScanFiles}
\title{Transfer files from all subfolders to analogous subfolders}
\usage{
transferScanFiles(from, to, askBeforeContinue = TRUE)
}
\arguments{
\item{from}{path to a folder whose subfolders you want checked,
(e.g. "C:/cg2015scans" or "I:/Departments/Research/EchinaceaCG2015/cg2015scans")}

\item{to}{the drive on which the wanted folder is found (e.g. "F:/", "D:/", or
"I:/Departments/Research/EchinaceaCG2015/")}

\item{askBeforeContinue}{logical: should a message be printed before continuing
to check other subfolders}
}
\description{
This function will user the transferFiles function to go through all
subfolders of the selected folder and transfer files not in one folder
to the other
}
\details{
Works from any from and any to as long as the name of the from folder
ends with "cgXXXXscans" where XXXX is a number
}
\examples{
\dontrun{
path1 <- "C:/cg2014scans"
pathos <- "E:/"
transferScanFiles(pathos, path)}

}
\keyword{scans}

