% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeAceProgress.R
\name{visualizeAceProgress}
\alias{visualizeAceProgress}
\title{Visualize ACE progress in the lab}
\usage{
visualizeAceProgress(experiment = "rem", year = "2020")
}
\arguments{
\item{experiment}{can be rem, cg, la, or lp}

\item{year}{the year the heads/pods were collected}
}
\value{
a barplot png showing the total number of heads/batches and the number that are done for each stage of the ACE workflow. 
The barplot includes the stages that are listed in the csv in Dropbox/visualizeAceProgress.
The dates above each bar are the dates when each stage was last updated.
}
\description{
Visualize ACE progress in the lab for rem, cg, la (Liatris aspera), or lp (Lilium philadelphicum).
Update the csv in Dropbox/visualizeAceProgress with the most recent counts of how many heads/batches are done.
If you are in a repo, run build.R or pathUpdate.R first so that dropboxPath is in the Global Environment.
If dropboxPath exists, a barplot png will be saved in Dropbox/visualizeAceProgress. If not, the png will be saved 
to the workspace.
}
