% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manageCountData.R
\name{writeUploadCSV}
\alias{writeUploadCSV}
\title{Create and write an upload csv}
\usage{
writeUploadCSV(scansFolder, writeTo, year, exprio, oneCt = NULL)
}
\arguments{
\item{scansFolder}{the folder in which the scans can be found}

\item{writeTo}{the file to which the upload csv will be written}

\item{year}{the harvest year of the scans}

\item{exprio}{a data frame containing two columns: (1) the experiments
to upload and (2) the priorities for the experiments. The first column
should have the experiments and the second column should have priorities}

\item{oneCt}{a data frame whose first column is experiments that should
only be counted once and whose second column is the username of the person
to count that experiment}
}
\value{
none, write to a csv
}
\description{
This will create and write a csv with the counting assignments as it
should be ready to upload
}

