% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echinaceaLab-package.R
\docType{package}
\name{echinaceaLab-package}
\alias{echinaceaLab}
\alias{echinaceaLab-package}
\title{echinaceaLab: useful functions for the Echinacea Project}
\description{
Functions to help manage and visualize Echinacea Project datasets.
}
\keyword{internal}
