\name{echinaceaLab-package}
\alias{echinaceaLab-package}
\alias{echinaceaLab}
\docType{package}
\title{
echinaceaLab
}
\description{
Functions to help manage and visualize Echinacea Project datasets.
}
\details{
\tabular{ll}{
Package: \tab echinaceaLab\cr
Type: \tab Package\cr
Version: \tab 0.100\cr
Date: \tab 2015-01-21\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Author: Stuart Wagenius <swagenius@chicagobotanic.org>
Maintainer: Stuart Wagenius <swagenius@chicagobotanic.org>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages
}
\examples{
standardizeTtColors("black") # simple examples of the most important functions
}
