% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massFunctions.R
\name{listBadFiles}
\alias{listBadFiles}
\title{Find mass files that don't make proper csvs,}
\usage{
listBadFiles(path = ".")
}
\arguments{
\item{path}{character designating directory containing files of interest}
}
\value{
character vector of bad file names
}
\description{
Find mass files that don't return a good csv files. Bad files usually result from an extra
comma in the first line.
}
\examples{

\dontrun{
setwd("I:\\\\Departments\\\\Research\\\\EchinaceaVolunteers\\\\Balance\\\\sampleForEchLab\\\\CG2009_rawFiles")
listBadFiles( )
}
}
\seealso{
\code{\link{combineMassFiles}} and \code{\link{readMassFile}} and
  \code{\link{investigateMassFiles}} which are other useful functions that
  deal with mass files
}
\keyword{file}
\keyword{mass}

